Building components using dynamic imports and lazy statements
