+++ docs/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html lang="en"><head><meta charset="utf-8"/><link rel="icon" href="/jargon-cards-v2/favicon.ico"/><meta name="viewport" content="width=device-width,initial-scale=1"/><meta name="theme-color" content="#000000"/><meta name="description" content="Web site created using create-react-app"/><link rel="apple-touch-icon" href="/jargon-cards-v2/logo192.png"/><link rel="manifest" href="/jargon-cards-v2/manifest.json"/><title>React App</title><script defer="defer" src="/jargon-cards-v2/static/js/main.95a4a04b.js"></script><link href="/jargon-cards-v2/static/css/main.073c9b0a.css" rel="stylesheet"></head><body><noscript>You need to enable JavaScript to run this app.</noscript><div id="root"></div></body></html>
+<!doctype html><html lang="en"><head><meta charset="utf-8"/><link rel="icon" href="/jargon-cards-v2/favicon.ico"/><meta name="viewport" content="width=device-width,initial-scale=1"/><meta name="theme-color" content="#000000"/><meta name="description" content="Web site created using create-react-app"/><link rel="apple-touch-icon" href="/jargon-cards-v2/logo192.png"/><link rel="manifest" href="/jargon-cards-v2/manifest.json"/><title>React App</title><script defer="defer" src="/jargon-cards-v2/static/js/main.ab785a4a.js"></script><link href="/jargon-cards-v2/static/css/main.073c9b0a.css" rel="stylesheet"></head><body><noscript>You need to enable JavaScript to run this app.</noscript><div id="root"></div></body></html>
