+++ docs/static/js/95.0fd4889b.chunk.js
@@ -1,0 +1,2 @@
+"use strict";(self.webpackChunkjargon_cards_v2=self.webpackChunkjargon_cards_v2||[]).push([[95],{95:function(e,s,r){r.r(s);var a=r(428),n=(r(508),r(791),r(87)),c=r(184);s.default=function(){return(0,c.jsx)("div",{className:"App",children:(0,c.jsxs)("header",{className:"App-header",children:[(0,c.jsx)("img",{src:a.Z,className:"App-logo",alt:"logo"}),(0,c.jsx)("p",{children:"This is the main page!"}),(0,c.jsxs)("p",{children:["Edit ",(0,c.jsx)("code",{children:"src/App.js"})," and save to reload."]}),(0,c.jsx)(n.rU,{to:"about",children:"About this page"}),(0,c.jsx)("a",{className:"App-link",href:"https://reactjs.org",target:"_blank",rel:"noopener noreferrer",children:"Learn React"})]})})}}}]);
+//# sourceMappingURL=95.0fd4889b.chunk.js.map
