+++ src/App.js
@@ -2,10 +2,12 @@ import logo from './logo.svg';
 import './App.css';
 
 import React, { Suspense, lazy } from 'react';
+
 // import { BrowserRouter as Router } from 'react-router-dom';
 import { HashRouter as Router } from 'react-router-dom';
 import { Routes, Route } from 'react-router-dom';
-import { Link } from "react-router-dom";
+
+const Home = lazy(() => import('./routes/Home'));
 
 const App = () => (
   <Router /* basename="/jargon-cards-v2" */ >
@@ -17,21 +19,6 @@ const App = () => (
     </Suspense>
   </Router>
 );
-
-function Home() {
-  return <div className="App">
-    <header className="App-header">
-      <img src={logo} className="App-logo" alt="logo" />
-      <p>
-        Edit <code>src/App.js</code> and save to reload.
-      </p>
-      <Link to="about">About this page</Link>
-      <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
-        Learn React
-      </a>
-    </header>
-  </div>;
-}
 
 function About(params) {
   return <div className="App">
